+++ apps/cli/CHANGELOG.md
@@ -77,3 +77,6 @@
 ### Patches
 
 - `-v` flag replaced with `-V`
+- Removed `keyshade secret get` command
+- Updated `keyshade secret update` and `keyshade secret create` commands to include `-e/--entry` flag for specifying values to environments
+- Formatted outputs of `keyshade secret list` and `keyshade secret revisions` commands